WIP: major rewrite: replace SQL with YAML database
CReMa is undergoing a major rewrite in which the SQL database is
replaced by a YAML file. This has the advantage of using plain text data
and thus allows much easier version controlling.

This first work has been the progress of gradual testing and rewriting
of the basic functions. Once the full functionality of CReMa as in the
SQL version has been achieved, this will be merged.

Note: this already adds bibtex import functionality.