Insert columns sorted by near front of list view
When sorting against a key, the columns used to be appended to the list
view. However, with long titles this often leads to very long outputs.
Instead, the columns sorted by are now inserted near the front of the
view which helps a lot with readability, especially in the new
curses-TUI.

When filtering against a key, the columns are still appended because the
filtered key is most likely of less interest here.

Fixes #10