Reset scrolling position when entering Show view
Fixes #13

When entering the Show view, the scrolling position needs to be reset.
Otherwise an inconsistent state can be reached because the buffer length
changes when entering this view.
When leaving the view, the scrolling position is set in such a way to
ensure the current line is visible. It defaults to placing it on the top
line.