Extract database handling into separate module
By extracting the database handling into its own module, it becomes
accessible from anywhere in the code without causing a cyclic dependency
(as it would if it were handled directly in the config module because it
depends on the parser which in turn requires the config).

Through this centralization it suffices to load the database once on
startup and whenever the database is written again. This improves the
performance of the TUI drastically!