Ensure python3 is used and respective $PATH settings
Although Python 2 was sunset this year, it and Python 3 may still
coexist on various platforms. This change ensures, that Python 3 is used
for installing.
Furthermore, the executables are ensured to respective the $PATH
settings and use the first Python 3 version specified there rather than
defaulting to the gobal executable in /usr/bin