Fix #48
Handle prompt string which exceed the current terminal width by
replacing the prompt with a curses pad (rather than window) and resizing
it when necessary.
This also allows us to handle a curses bug in which the `addstr` method
fails when given text containing newline-characters. In these cases we
also provide a popup with the full message and print its first line to
the prompt to provide context on the last message once the popup has
been closed again.

Note: due to the switch from the window to a pad we will not be able to
make use of curses' Textboxes which could have been an alternative to
the manual text input handling which we have at the moment.
[Mentioning this, just to note this down somewhere...]