Allow entry addition through the Edit command
We can leverage the Edit command to allow manual entry addition when the
label to be edited could not be found.
In order to ensure this is what the user really wants to do, we require
an additional command line argument for this scenario to happen. If it
is not supplied, the user will be informed by the logger.