Implement #45
This adds a basic ISBN parser based on the open-access (and keyless) API
provided by openlibrary.org
The parser is quite minimal and only tries to extract a very crude set
of data to populate the entry with. More detailed information should be
added by the user manually but this is nonetheless a better starting
point than having to do everything manually.