Fix #50
Rather than a literal manual search for ANSI color codes during viewing
a TextBuffer we can use a regex pattern to detect the escape sequences
more consistently and *in order*. This allows us to use a stack-like
structure to store the spanned regions for each color.
By sorting the colors before applying them in the curses API we can
respect the color priorities, too.

Note, that we also change the `\033[` escape code to `\x1b[` as this
appears to be the raw code which the regex library returns. Although
this is no longer POSIX compliant I cannot think of a use-case where
this would matter.