Handle invalid arXiv and DOI IDs gracefully.
Fixes #57

This should have already been addressed when #53 surfaced. With this
commit, all parsers which rely on loading the bibliography data through
an API can handle possible errors gracefully.
(Actually, the DOI case is asserted prior to actually querying the data
and should not run into such an error. But this commit ensures
ReadTimeout error are caught. I also increased the hard-coded timeout to
reduce the amount of times this happens during testing.)

Furthermore, unittests are added for each of these scenarios to prevent
regressions.