Deprecate init's --force argument.
I don't think there is any benefit to providing the user the option to
nuke their database file from the CLI. Especially, since this operation
will rarely be performed and editing the file manually is just as easy.

Furthermore, I plan to add Git integration in the next release, which is
going to add the `--git` argument (abbreviated by `-g`). This option
will support safe "re-initialization" in order to start tracking an
existing literature file, too. Imagine the problems that could arise if
a user accidentally mistyped `cobib init -f` rather than `cobib init -g`
which is not unlikely to happen if anyone else out there types similar
to me.