Add basic git integration.
CoBib can be configured to use git to track the history of the database file. For now, this will result in simple auto-generated commits for each command which affects the database file. In the future, more advanced configuration options such as syncing with remotes and direct history-affecting commands such as Undo and Redo may be added.