Use yaml's old API for loading
As described here [1], the new API is performing slightly worse than the
old one. Since we are not really taking advantage of any parts of the
new API here, reverting to the old one boosts performance for loading
the database file. This is more noticeable the larger the database is.

[1] https://sourceforge.net/p/ruamel-yaml/tickets/159/#8f94