Serve the documentation at https://mrossinek.gitlab.io/cobib
This updates the CI to automatically deploy the documentation (generated by [pdoc](https://pdoc3.github.io/pdoc/)) through GitLab pages.
Every branch will serve a local version of the documentation in the `dev` environment which can be explored through the `View App` button from within a merge request.

When a commit is pushed to `master`, the `pages` job can be triggered manually which will update the online documentation. This should only be done with each release of CoBib.