Fix #73
With pdoc 6.4.3, the issue hindering the inclusion of the `tests` module
in the documentation has been resolved.
This release also updated the `index.html` template allowing to remove a
lot of the duplicated information from coBib's custom version.

The actual documentation contents of the `tests` module will have to be
updated later.