Fix type annotation in docs
pdoc is able to resolve forward references if the module can be resolved
properly. It cannot leverage TYPE_CHECKING imports because it is not a
static scenario (because it actually imports everything).
Changing the conditional imports to a higher level allows pdoc to do the
type resolution *and* have correct type annotations (although more
verbose which is really not a problem given the pdoc benefits).