Fix metadata querying
The `PKG-INFO` metadata is a wrapper of `.egg-info` files, which are
used by editable installs (which is why my local testing of
fc4aac3debc5e180afc98955ef111f6b6fb600bf was working fine).
However, in general users will not use an editable install in which case
the `METADATA` file must be used instead.

This will likely all be simplified once Python 3.10 includes the fix for
[this bug](https://bugs.python.org/issue43926).