breaking: move cobib.commands.list to cobib.commands.list_
Having a module name clashing with a python language keyword seems like
a call for trouble. This is the same reason why the `ImportCommand` is
located in the `import_` module.

Although technically a breaking change, I don't think anyone will really
notice. The bigger change which renamed the location of the
configuration settings (matching this module location) was already
applied and merged anyways.