fix: ensure UTF-8 encoding of downloaded data
I have had many encounters where the escaping of special characters
resulted in very odd output such as `{\~A}{\textparagraph}` which
originated from an `Ã¶` input. I only now realized that this was due to
the lack of a proper encoding being specified on the level of the HTTP
request! Ensuring a UTF-8 encoding is set, the input will be the proper
`ö` which will properly be escaped as `{\"o}`.