Refactors the `Popup` and `Prompt` UI components
* doc: add issue/MR references in Changelog

* Documentation updates

* Deprecate the `console` and `prompt` arguments to BaseCommand

* refactor: remove PopupPanel

* ci: enable pyflakes checks (to catch unused imports)

* Linting

* refactor: update add and open commands to use new Prompt

* refactor: initial extraction of the Prompt widget

This starts the extraction of the Prompt widget in a similar manner to
the Progress widget done a while ago. This will have the benefit of
simplifying its behavior because it removes UI-specific handling from
the responsibility of the commands (which should never have been put
there in the first place).
In this current form, this commit still breaks the `add` and `open`
commands, but these will be fixed shortly.

* refactor: mount the Progress indicators directly into the App

* refactor: use bulitin notifications for printing

* feat: automatically open the log screen upon ERROR messages

* refactor: remove the Popup class

* feat: render log messages on separate screen