Implements the new `review` command
* feat: implement an experimental in-line editing mode

* feat: implement the context argument and action

* feat: implement the fields filter

* refactor: implement the Entry.merge method

* tests: graceful handling of faulty resume SHAs

* feat: keep references to the done entries

* feat: implement the --resume option

* docs: Add MR references to the changelog

* Update man-page

* tests: more review command unittests

* tests: some basic review unittests

* docs: document the new review command

* feat: mostly implements the actual review command logic

* feat: add string-support to YAMLParse.parse

Rather than always interpreting the provided argument as the path to a
file, it can now also handle strings which directly contain YAML data.

* refactor: add `cobib.utils.context.get_active_app`

This method uses the `contextvars` module and extract the textual App
which registers itself in the current `Context` to extract it. This
avoids the TUI having to register itself in all of these utility classes
as noted in the Changelog.

This also refactors the `EditCommand` and extracts the editing method
into a `staticmethod` which makes it reusable by the `ReviewCommand`.

* feat: continue the implementation of the review command

* fix: ensure help is compatible with pre_prompt

* fix: MainContent handling in TUI

Whenever the list of entries gets updated from a callback while the
LogScreen is current, this call would fail. My previous attempt at
fixing did not quite identify the actual problem, which is now fixed by
this commit.

* fix: some minor TUI+prompt interaction fixes

* WIP: adds the basic structure for the review command

* feat: ensuring the '--' pseudo-argument works more consistently