add entry-points
* docs: update man-page

* docs: update Changelog

* docs: update dummy-plugin reference in main README

* fix: additional module-level imports

* plugin: docs: additional documentation

* plugin: test: use porcelain mode

* plugin: fix: add missing testing database

* plugin: some basic unittests

* refactor: relocate testing databases

* refactor: clean up Command interface

This removes the deprecated `console` and `prompt` attributes from all
commands.

* ci: update pytest to latest version

Refactoring the shell helpers and their unittests resolved the problems
I was having with pytest>=8

Closes #137

* refactor: LintCommand

* tests: skip tests pending refactoring

* refactor: UnifyLabelsCommand

* refactor: move ExampleConfigCommand to the config module

* refactor: ExampleConfigCommand

* refactor: remove special handling of init command

This also removes the initialization of the Database from the CLI level
in favor of loading it only on-demand during the command execution
(which works because the Database is implemented as a Singleton).

* refactor: remove ShellHelper UI

* refactor: remove deprecated shell helper functions

* Linting

* fix: forward-compatibility to Python 3.12

* Add a dummy plugin

* refactor: use entry-points for command, importers, and parsers

This refactors these modular aspects of coBib as entry-points, allowing
them to be extended by plugins.