fix: support database locations in nested folders
* fix: support database locations in nested folders

Up until now, the git integration did not support placing the database
file in nested folders of the git repository. The reason for this was
the hard-coded check for the `.git` folder to be in the same folder as
the database file itself.

This commit fixes this by making the check for git integration more
rigorous.