docs: add the man-page section indicator to HTML
When viewing the online documentation (which now includes rendered
man-pages, it is not apparent from which man-page section the output is
coming from, since this information is usually not shown in the .NAME
section but rather the title (which we omit on the HTML fragment).

However, it would be useful for readers of the online documentation to
have this information if they choose to want to look up the man-page in
the future. Thus, this commit inserts the section indicator into the
name.