test: refactor help notification test
The flakiness of the test was not limited to specific Python versions as
initially thought. The problem appears to be minor fluctuations in the
type-setting of line breaks in textual's notification popups.

This commit replaces the snapshot test with a more simplistic test that
asserts the notification has occurred but does not test its visual
appearance.