Implement an interactive shell
* test: fix tests for arbitrary terminal size

* test: increase Shell testing

* fix: do not flush a possibly closed handler

* fix: ensure Python 3.9 compatibility

I am not entirely sure why, but the unconditional removal of
LoggingHandler instances caused issues with Python 3.9.
I originally introduced this removal to avoid interactions of the Shell
and TUI tests. Now, this fix attempts to ensure that the TUI tests
within themselves do not break under this Python version.

* docs: update changelog

* fix: also remove previously registered LoggingHandlers

This is necessary because these handlers can accumulate during the test
suite.

* fix: test inside tox

* test: the Shell's help alias

* fix: very minor tweaks

* test: basic Shell trigger

* test: remove duplicate test

* test: update snapshots with new man-page contents

* test: update snapshots with new log message styling

* docs: warn about ignoring porcelain mode in interactive UIs

* feat: support help/exit/quit in Shell

* refactor: reduce code duplication

* fix: reduce default log level in the shell

* docs: start to document the new shell UI

* fix: remove log time in Shell

* refactor: make log messages less glaring in UIs

* refactor: move console to UI components

* refactor: generalize LoggingHandler to support more UIs

* refactor: integrate LoggingHandler into Shell

In fact, we do want to subclass the UI... We already solved the
duplicate logging handler problem in the TUI... Next, we need to
de-duplicate the LoggingHandler logic.

* fix: do not subclass UI in Shell

There is no reason to do so, since none of the functions are needed and
`UI.load_command` is already a staticmethod. Furthermore, this fixes a
problem where the UI's __init__ method registers a log handler with the
ROOT logger multiple times.

* refactor: centralize rich.Console instance

* test: update manpage index snapshots after adding new man-page

* [wip] feat: an interactive shell

* refactor: de-duplicate command loading logic