feat: add an interactive tutorial
* docs: update tutorial instructions after final readthrough

* docs: de-duplicate module names

* docs: include editing URLs for other modules

* docs: update changelog

* refactor: de-duplicate mock_stdin logic

* test: minimal testing of the tutorial

* test: update snapshots

* fix: patch MockStdin.fileno

* docs: update tutorial instructions for open

* docs: for TUTORIAL_* file paths

* docs: minor tweaks to the tutorial instructions

* fix: import bibtex tests

* docs: finish the tutorial instructions

* refactor: always insert horizontal ruler before tutorial prompt

* [wip] docs: continue the tutorial instructions

* fix: parse Shell inputs through shlex

* [wip] docs: continue the tutorial instructions

* fix: lint

* [wip] docs: continue the tutorial instructions

* fix: circular import

* fix: URL shifts with new basename

* fix: install_man_pages target

* fix: detect if man-page targets are up-to-date

* meta: extract `make docs` target

* fix: detect if html_fragment targets are up-to-date

This avoids re-generating all man-page HTML fragments every single time.

* meta: allow specifying which module docs to re-render

* docs: include preview of tutorial instructions in online docs

This was the main motivation for switching the API docs generation from
pdoc's CLI to API. Basically, we can preview the markdown-based tutorial
instructions in the online docs by injecting them as the Enum member
docstrings. They cannot be implemented that way directly, because Python
does not have docstrings on variables and this is only an inofficial
feature supported by many API documentation generators. And the preview
of the Enum value does not render nicely, hence this injection.

* meta: replace vim magic with updates during docs/make.py

Rather than rely on vim to be present, we can now simply fix the ToC of
the main module page from within python.

* meta: switch pdoc to library-based call

This changes how we build the API docs. Rather than rely on pdoc's basic
CLI, we now have a basic Python script and leverage its API to generate
the docs. This will allow us to make fine-tuned adjustments more easily.

* [wip] docs: write the tutorial instructions

* docs: complete TutorialCommand API docs

* docs: complete cobib-tutorial.1

* refactor: move testing import database for tutorial usage

* fix: disallow tutorial command within Shell and TUI

* docs: update cobib-getting-started.7

* docs: prepare the cobib-tutorial.1 man page

* [wip] feat: basic interactive tutorial skeleton